requirements.txt
streamlit
pypdf2
fpdf2
google-generativeai
pymupdf